requirements.txt
pillow==12.1.1
stepic==0.5.0
tkinterdnd2==0.3.0